requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
Click==7.0
Django==3.0.3
django-crispy-forms==1.9.0
freeze==1.0.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.7.0
pylint==2.4.4
pytz==2019.3
regex==2020.2.20
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
